fix: rename kms to cipher
