+++ src/lib.rs
@@ -15,11 +15,11 @@
 
 #![feature(btree_cursors)]
 
+pub mod cipher;
 pub mod cli;
 pub mod common;
 pub mod handler;
 pub mod job;
-pub mod kms;
 pub mod router;
 pub mod service;
